requirements.txt
Django==1.11.3
Pillow==4.2.1
boto==2.48.0
boto3==1.4.4
botocore==1.5.80
certifi==2017.4.17
chardet==3.0.4
django-storages==1.6.3
django-widget-tweaks==1.4.1
djangorestframework==3.6.3
docutils==0.13.1
futures==3.1.1
idna==2.5
jmespath==0.9.3
olefile==0.44
psycopg2==2.7.1
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.1
s3transfer==0.1.10
six==1.10.0
urllib3==1.21.1
wsgiref==0.1.2
